// frontend/src/app/model/zaposleni.model.ts
export class Zaposleni {
    username : string = "";
    password:  string = "";
    ime:  string = "";
    prezime:  string = "";
    adresa: string = "";
    telefon:  string = "";
    vebsajt:  string = "";
    podaci:  string = "";
    zvanje: string = "";
    kabinet:  string = "";
    status:  string = "";
    slika: string = "";
}